+++ framework/src/cli/config/factory.ts
@@ -1,25 +1,15 @@
 import Hashmap from '../../hashmap.js';
 import {
-  CLI_OPTION_MIN_LENGTH,
   DEFAULT_REPETITIONS,
   DEFAULT_THREADS,
-  FIRST, FIRST_ARGUMENT,
-  SECOND,
-  TWO,
-  FIVE,
+  FIRST_ARGUMENT,
 } from '../../constants.js';
 import Config from './config.js';
-import {
-  camelCase,
-} from 'change-case';
-import toValue from './to-value.js';
+import fromEnv from './from-env.js';
+import fromCli from './from-cli.js';
 
 // eslint-disable-next-line complexity
 export default (cwd: string, args: string[], env: Hashmap,) => {
-  const options = args.filter(
-    (option,) => option.startsWith('--',)
-      && option.length >= CLI_OPTION_MIN_LENGTH,
-  );
   const config: Config = {
     cwd,
     threads: DEFAULT_THREADS,
@@ -28,16 +18,7 @@ export default (cwd: string, args: string[], env: Hashmap,) => {
       (arg,) => ! arg.startsWith('--',),
     )[FIRST_ARGUMENT] || 'help',
   };
-  for (const key of Object.keys(env,)) {
-    if (key.startsWith('IAB_',)) {
-      config[camelCase(key.substring(FIVE,),)] = toValue(env[key],);
-    }
-  }
-  for (const option of options) {
-    const parts = option.split('=',);
-    if (parts.length === TWO) {
-      config[parts[FIRST]] = toValue(parts[SECOND],);
-    }
-  }
+  fromEnv(config, env,);
+  fromCli(config, args,);
   return config;
 };
